requirements.txt
streamlit
pytesseract
pdf2image
Pillow
scikit-learn
pandas
numpy
joblib
google-generativeai
python-dotenv